chore: run e2e tests with testcontainers
This change allows for the use of testcontainers to run E2E tests. Why
would we want this? Mainly for three reasons:

1. IDE friendlier
2. Debugger friendlier
3. Much easier to setup

It's a "one click" way (or a simple `go test e2e` way) to run E2E tests,
set the environment variable `ARGOCD_E2E_USE_TESTCONTAINERS=true` and
you're off to the races.

This has a bunch of changes, but the majority of them is simply creating
builders for each of the affected controller commands to separate builders,
which are used to be able instantiate each controller programatically.
The commands make use of these builders to set the flags as they
previously did.

Redis, K3S and the E2E server are run as testcontainers when running the
E2E tests. The controllers are run in-process but spun off into separate
goroutines.

Signed-off-by: Blake Pettersson <blake.pettersson@gmail.com>